requirements.txt
firecrawl-py
python-dotenv
pandas
groq
langchain-groq
langchain_google_genai
langchain
streamlit
openpyxl
faiss-cpu
groq
langchain-groq
pyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf